% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexContour.R
\name{meanderingTriangles}
\alias{meanderingTriangles}
\title{Meandering triangles for hexagonal grid in C++}
\usage{
meanderingTriangles(x.coords.left, x.coords.right, y.coords, z, levels)
}
\arguments{
\item{x.coords.left}{Vector for x coords of left-aligned rows (row 1,3,5,...)}

\item{x.coords.right}{Vector for x coords of right-aligned rows (row 2,4,6,...)}

\item{y.coords}{Vector for y coords of all rows.}

\item{z}{Matrix for elevation values for the grid point}

\item{levels}{Vector of z value cutoffs for contouring.}
}
\value{
list of x, y, and ID, for the contour line at each levels.
}
\description{
Meandering triangles for hexagonal grid in C++
}
\details{
This function is not meant to be used as is, unless you are very familiar 
with how hexContour works.
}
\references{
Hill, B. (2017) Meandering triangles. Naming Things. 
https://blog.bruce-hill.com/meandering-triangles
}
